+++ hata/discord/audit_logs/audit_log_entry_change_conversions/onboarding_screen.py
@@ -1,11 +1,22 @@
 __all__ = ()
 
-from ...onboarding import OnboardingPrompt
-from ...onboarding.onboarding_screen.fields import validate_default_channel_ids, validate_enabled, validate_prompts
+from ...onboarding import OnboardingMode, OnboardingPrompt
+from ...onboarding.onboarding_screen.fields import (
+    validate_default_channel_ids, validate_enabled, validate_mode, validate_prompts
+)
 
-from ..audit_log_change.flags import FLAG_IS_MODIFICATION
+from ..audit_log_change.flags import FLAG_IS_IGNORED, FLAG_IS_MODIFICATION
 from ..audit_log_entry_change_conversion import AuditLogEntryChangeConversion, AuditLogEntryChangeConversionGroup
 from ..conversion_helpers.converters import get_converter_ids, put_converter_ids
+
+
+# ---- below_requirements ---- 
+
+BELOW_REQUIREMENTS_CONVERSION_IGNORED = AuditLogEntryChangeConversion(
+    'below_requirements',
+    '',
+    FLAG_IS_IGNORED,
+)
 
 
 # ---- default_channel_ids ----
@@ -18,6 +29,7 @@ DEFAULT_CHANNEL_IDS_CONVERSION = AuditLogEntryChangeConversion(
     put_converter = put_converter_ids,
     validator = validate_default_channel_ids,
 )
+
 
 # ---- enabled ----
 
@@ -34,6 +46,27 @@ def enabled_get_converter(value):
     if value is None:
         value = True
     return value
+
+
+# ---- mode ----
+
+MODE_CONVERSION = AuditLogEntryChangeConversion(
+    'mode',
+    'mode',
+    FLAG_IS_MODIFICATION,
+    validator = validate_mode,
+)
+
+
+@MODE_CONVERSION.set_get_converter
+def mode_get_converter(value):
+    return OnboardingMode.get(value)
+
+
+@MODE_CONVERSION.set_put_converter
+def mode_put_converter(value):
+    return value.value
+
 
 
 # ---- prompts ----
@@ -71,7 +104,9 @@ def prompts_put_converter(value):
 # ---- Construct ----
 
 ONBOARDING_SCREEN_CONVERSIONS = AuditLogEntryChangeConversionGroup(
+    BELOW_REQUIREMENTS_CONVERSION_IGNORED,
     DEFAULT_CHANNEL_IDS_CONVERSION,
     ENABLED_CONVERSION,
+    MODE_CONVERSION,
     PROMPTS_CONVERSION,
 )