+++ hata/discord/audit_logs/audit_log_entry_change_conversions/soundboard_sound.py
@@ -1,11 +1,13 @@
 __all__ = ()
 
-from ...emoji import create_partial_emoji_data, create_partial_emoji_from_data
-from ...soundboard.soundboard_sound.fields import validate_available, validate_emoji, validate_name, validate_volume
 
-from ..audit_log_change.flags import FLAG_IS_MODIFICATION
+from ...soundboard.soundboard_sound.fields import (
+    validate_available, validate_user_id, validate_id, validate_name, validate_volume
+)
+
+from ..audit_log_change.flags import FLAG_IS_IGNORED, FLAG_IS_MODIFICATION
 from ..audit_log_entry_change_conversion import AuditLogEntryChangeConversion, AuditLogEntryChangeConversionGroup
-from ..conversion_helpers.converters import get_converter_name, put_converter_name
+from ..conversion_helpers.converters import get_converter_id, get_converter_name, put_converter_id, put_converter_name
 
 
 # ---- available ----
@@ -26,30 +28,39 @@ def available_get_converter(value):
 
 
 # ---- emoji ----
+# Need to add support for this as well
+# EMOJI_CONVERSION_0 = AuditLogEntryChangeConversion(
+#     'emoji_id',
+#     'emoji',
+#     FLAG_IS_MODIFICATION,
+#     validator = validate_emoji,
+# )
+#
+#
+# EMOJI_CONVERSION1 = AuditLogEntryChangeConversion(
+#     'emoji_name',
+#     'emoji',
+#     FLAG_IS_MODIFICATION,
+#     validator = validate_emoji,
+# )
 
-EMOJI_CONVERSION = AuditLogEntryChangeConversion(
-    'emoji',
-    'emoji',
+# ---- id  ----
+
+ID_CONVERSION = AuditLogEntryChangeConversion(
+    'sound_id',
+    'id',
     FLAG_IS_MODIFICATION,
-    validator = validate_emoji,
+    get_converter = get_converter_id,
+    put_converter = put_converter_id,
+    validator = validate_id,
 )
 
 
-@EMOJI_CONVERSION.set_get_converter
-def emoji_get_converter(value):
-    if (value is not None):
-        value = create_partial_emoji_from_data(value)
-    
-    return value
-
-
-@EMOJI_CONVERSION.set_put_converter
-def emoji_put_converter(value):
-    if value is not None:
-        value = create_partial_emoji_data(value)
-    
-    return value
-
+ID_CONVERSION_IGNORED = AuditLogEntryChangeConversion(
+    'id',
+    '',
+    FLAG_IS_IGNORED,
+)
 
 # ---- name ----
 
@@ -60,6 +71,17 @@ NAME_CONVERSION = AuditLogEntryChangeConversion(
     get_converter = get_converter_name,
     put_converter = put_converter_name,
     validator = validate_name,
+)
+
+# ---- user_id ----
+
+USER_ID_CONVERSION = AuditLogEntryChangeConversion(
+    'user_id',
+    'user_id',
+    FLAG_IS_MODIFICATION,
+    get_converter = get_converter_id,
+    put_converter = put_converter_id,
+    validator = validate_user_id,
 )
 
 
@@ -85,7 +107,9 @@ def volume_get_converter(value):
 
 SOUNDBOARD_SOUND_CONVERSIONS = AuditLogEntryChangeConversionGroup(
     AVAILABLE_CONVERSION,
-    EMOJI_CONVERSION,
+    ID_CONVERSION,
+    ID_CONVERSION_IGNORED,
     NAME_CONVERSION,
+    USER_ID_CONVERSION,
     VOLUME_CONVERSION,
 )