+++ hata/discord/audit_logs/audit_log_entry_detail_conversions/application_command.py
@@ -1,6 +1,6 @@
 __all__ = ()
 
-from ...application_command.application_command_permission.fields import validate_application_id
+from ...application_command.application_command_permission.fields import validate_application_id, validate_guild_id
 from ...application_command.application_command_permission_overwrite.fields import validate_channel_id
 
 from ..audit_log_entry_detail_conversion import AuditLogEntryDetailConversion, AuditLogEntryDetailConversionGroup
@@ -17,6 +17,7 @@ APPLICATION_ID_CONVERSION = AuditLogEntryDetailConversion(
     validator = validate_application_id,
 )
 
+
 # ---- channel_id ----
 
 CHANNEL_ID_CONVERSION = AuditLogEntryDetailConversion(
@@ -28,9 +29,21 @@ CHANNEL_ID_CONVERSION = AuditLogEntryDetailConversion(
 )
 
 
+# ---- guild_id ----
+
+GUILD_ID_CONVERSION = AuditLogEntryDetailConversion(
+    'guild_id',
+    'guild_id',
+    get_converter = get_converter_id,
+    put_converter = put_converter_id,
+    validator = validate_guild_id,
+)
+
+
 # ---- Construct ----
 
 APPLICATION_COMMAND_CONVERSIONS = AuditLogEntryDetailConversionGroup(
     APPLICATION_ID_CONVERSION,
     CHANNEL_ID_CONVERSION,
+    GUILD_ID_CONVERSION,
 )