+++ hata/discord/audit_logs/audit_log_entry_detail_conversions/tests/test__application_command.py
@@ -1,11 +1,13 @@
 import vampytest
 
-from ....application_command.application_command_permission.fields import validate_application_id
+from ....application_command.application_command_permission.fields import validate_application_id, validate_guild_id
 from ....application_command.application_command_permission_overwrite.fields import validate_channel_id
 
 from ...conversion_helpers.converters import get_converter_id, put_converter_id
 
-from ..application_command import APPLICATION_COMMAND_CONVERSIONS, APPLICATION_ID_CONVERSION, CHANNEL_ID_CONVERSION
+from ..application_command import (
+    APPLICATION_COMMAND_CONVERSIONS, APPLICATION_ID_CONVERSION, CHANNEL_ID_CONVERSION, GUILD_ID_CONVERSION
+)
 
 
 def test__APPLICATION_COMMAND_CONVERSIONS():
@@ -14,7 +16,7 @@ def test__APPLICATION_COMMAND_CONVERSIONS():
     """
     vampytest.assert_eq(
         {*APPLICATION_COMMAND_CONVERSIONS.get_converters.keys()},
-        {'application_id', 'channel_id'},
+        {'application_id', 'channel_id', 'guild_id'},
     )
 
 
@@ -38,3 +40,14 @@ def test__CHANNEL_ID_CONVERSION__generic():
     vampytest.assert_is(CHANNEL_ID_CONVERSION.get_converter, get_converter_id)
     vampytest.assert_is(CHANNEL_ID_CONVERSION.put_converter, put_converter_id)
     vampytest.assert_is(CHANNEL_ID_CONVERSION.validator, validate_channel_id)
+
+
+# ---- guild_id ----
+
+def test__GUILD_ID_CONVERSION__generic():
+    """
+    Tests whether ``GUILD_ID_CONVERSION`` works as intended.
+    """
+    vampytest.assert_is(GUILD_ID_CONVERSION.get_converter, get_converter_id)
+    vampytest.assert_is(GUILD_ID_CONVERSION.put_converter, put_converter_id)
+    vampytest.assert_is(GUILD_ID_CONVERSION.validator, validate_guild_id)