+++ hata/discord/audit_logs/audit_log_entry_detail_conversions/user.py
@@ -2,8 +2,21 @@ __all__ = ()
 
 from ...integration import IntegrationType
 from ...integration.integration.fields import validate_type as validate_integration_type
+from ...user.voice_state.fields import validate_channel_id
 
 from ..audit_log_entry_detail_conversion import AuditLogEntryDetailConversion, AuditLogEntryDetailConversionGroup
+from ..conversion_helpers.converters import get_converter_id, put_converter_id
+
+
+# ---- channel_id ----
+
+CHANNEL_ID_CONVERSION = AuditLogEntryDetailConversion(
+    'channel_id',
+    'channel_id',
+    get_converter = get_converter_id,
+    put_converter = put_converter_id,
+    validator = validate_channel_id,
+)
 
 
 # ---- count ----
@@ -91,6 +104,7 @@ def integration_type_put_converter(value):
 # ---- Construct ----
 
 USER_CONVERSIONS = AuditLogEntryDetailConversionGroup(
+    CHANNEL_ID_CONVERSION,
     COUNT_CONVERSION,
     DELETE_MESSAGE_DURATION_CONVERSION,
     INTEGRATION_TYPE_CONVERSION,